+++ hbw/app/javascript/packs/hbw.js
@@ -1,4 +1,5 @@
 import './hbw/init/dependencies';
+import './hbw/init/fontawesome';
 import './hbw/components';
 import './hbw/connection';
 import './hbw/dispatcher';