+++ CHANGELOG.md
@@ -49,6 +49,7 @@ v2.6.14 [unreleased]
 -------------------
 ### Bugfixes
 - [#601](https://github.com/latera/homs/pull/601) Move fontawesome to public/assets.
+- [#603](https://github.com/latera/homs/pull/603) Don't highlight multi nullable select_table.
 
 v2.6.13 [2021-09-03]
 -------------------